requirements.txt
Flask==3.0.3
kafka-python==2.2.13
redis==6.2.0